Resolve race condition fetching local resources
If fetching a local file completed too quickly, we migh decrement the
pending counter and abort the processing loop assuming all resources had
been processed. Instead we needed to decrement the pending count after
calling the handler so that the loop remains active until the resource
is ready to process.